Moved classes around and improved visibility
